requirements.txt
requests>=2.31.0
pandas>=2.1.0
pyarrow>=14.0.0
python-dateutil>=2.8.2
duckdb>=0.10.0
urllib3>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pyyaml>=6.0.0
colorlog>=6.7.0


